+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added fix for plan focus causing errorred GUI textures
 - Added tweak disabling item interaction on boiling Crucibles
 - Added tweak blacklisting entities from spreading Flux Phage
+- Added tweak disabling Thaumonomicon category zoom
 
 ---